+++ src/components/ui/mode-selector/index.tsx
@@ -1,2 +1,1 @@
-export { ModeSelector } from './mode-selector'
-export type { ModeSelectorProps } from './types'
+export { ModeSelector, type ModeSelectorProps } from './mode-selector'
