--- src/components/ui/mode-selector/types.ts
@@ -1,7 +1,0 @@
-import type { Mode } from '@/types'
-
-export type ModeSelectorProps = {
-  modes: Mode[]
-  selectedMode: Mode | null
-  onModeChange: (modeId: string) => void
-}
